Move tx.rs to the SDK crate
We also update all imports to properly refer to the new tx location
