+++ apps/src/lib/cli/client.rs
@@ -1,7 +1,8 @@
 use color_eyre::eyre::{eyre, Report, Result};
 use namada::ledger::eth_bridge::bridge_pool;
-use namada::ledger::tx::dump_tx;
-use namada::ledger::{signing, tx as sdk_tx};
+use namada::ledger::signing;
+use namada::sdk::tx as sdk_tx;
+use namada::sdk::tx::dump_tx;
 use namada::types::control_flow::ProceedOrElse;
 
 use crate::cli;
