+++ shared/src/ledger/masp.rs
@@ -62,10 +62,10 @@ use thiserror::Error;
 use crate::ledger::args;
 use crate::ledger::args::InputAmount;
 use crate::ledger::queries::Client;
-use crate::ledger::tx::decode_component;
 use crate::proto::Tx;
 use crate::sdk::rpc;
 use crate::sdk::rpc::{query_conversion, query_storage_value};
+use crate::sdk::tx::decode_component;
 use crate::tendermint_rpc::query::Query;
 use crate::tendermint_rpc::Order;
 use crate::types::address::{masp, Address};
