+++ shared/src/sdk/mod.rs
@@ -1,2 +1,5 @@
 //! Namada's SDK API
 pub mod rpc;
+
+#[allow(clippy::result_large_err)]
+pub mod tx;
